Core: Patch Drupal Core for PHP 7.4 compatability
When saving a Page node a PHP notice is raised: Trying to access
array offset on value of type null in
FieldInfo->prepareInstanceDisplay()

For some reason this occurs because the formatters for a relation
endpoint is left out deep down in the call stack. It may be related
to static variables and cache clearing in the process.

It should be OK to make the code conditional.

See
https://www.drupal.org/project/drupal/issues/2782809